[INLONG-10436][Manager] Move schedule configuration from stream to group (#10437)
