+++ inlong-manager/manager-schedule/src/main/java/org/apache/inlong/manager/schedule/ScheduleType.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.schedule;
+package org.apache.inlong.manager.schedule;
 
 import lombok.Getter;
