+++ inlong-manager/manager-service/src/test/java/org/apache/inlong/manager/service/workflow/group/CreateGroupWorkflowDefinitionTest.java
@@ -40,7 +40,8 @@ public class CreateGroupWorkflowDefinitionTest extends ServiceBaseTest {
         Assertions.assertNotSame(cloneProcess2, cloneProcess1);
         Assertions.assertNotNull(process.getTaskByName("InitMQ"));
         Assertions.assertNotNull(process.getTaskByName("InitSort"));
-        Assertions.assertEquals(2, process.getNameToTaskMap().size());
+        Assertions.assertNotNull(process.getTaskByName("InitSchedule"));
+        Assertions.assertEquals(3, process.getNameToTaskMap().size());
     }
 
 }
